pyproject.toml
[project]
name = "transactions_app"
version = "1.0"
requires-python = ">=3.8"
dependencies = [
  "grpcio==1.59.3",
  "grpcio-tools==1.59.3",
  "pydantic==2.5.2",
  "SQLAlchemy==2.0.23",
  "psycopg2-binary==2.9.9",
  "black==23.11.0",
  "pytest==7.4.3",
]

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  src/grpc_output
)/
'''


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
